+++ packages/antd-management-fast-framework/src/framework/DataListView/BatchAction/index.jsx
@@ -1,79 +1,111 @@
 import { Button, Dropdown } from 'antd';
-import { ConfigConsumer } from 'antd/lib/config-provider/context';
-import React from 'react';
+import React, { PureComponent } from 'react';
 
-import { AnchorLink, iconBuilder } from 'antd-management-fast-component';
+import { connect } from 'easy-soft-dva';
+import { isFunction } from 'easy-soft-utility';
+
+import { IconInfo } from 'antd-management-fast-component';
+
+import { switchControlAssist } from '../../../utils';
 
 import styles from './index.less';
 
-/**
- * 默认的 index 列容器，提供一个好看的 index
- * @param param0
- */
-const DropdownButton = ({
-  children,
-  menus = [],
-  onSelect,
-  style,
-  disabled,
-}) => (
-  <ConfigConsumer>
-    {() => {
-      const items = menus.map((item) => {
-        return {
-          key: item.key,
-          label: item.name,
-          icon: item.icon,
-          disabled: (item.disabled || null) == null ? false : item.disabled,
-        };
-      });
+@connect(({ switchControl }) => ({
+  switchControl,
+}))
+class BatchAction extends PureComponent {
+  render() {
+    const {
+      switchControl,
+      flag,
+      type,
+      icon,
+      text,
+      size,
+      danger,
+      menus,
+      onClick,
+      onSelect,
+      disabled,
+    } = this.props;
 
+    const loading = switchControlAssist.check(switchControl, flag);
+
+    const items = menus.map((item) => {
+      return {
+        key: item.key,
+        label: item.name,
+        icon: item.icon,
+        disabled: (item.disabled || null) == null ? false : item.disabled,
+      };
+    });
+
+    if (items.length === 0) {
       return (
-        <Dropdown
+        <Button
+          type={type}
+          size={size}
+          disabled={disabled || loading}
+          danger={danger}
+          onClick={() => {
+            if (!isFunction(onClick)) {
+              return;
+            }
+
+            onClick();
+          }}
+        >
+          <IconInfo icon={icon} text={text} />
+        </Button>
+      );
+    }
+
+    if (isFunction(onClick)) {
+      return (
+        <Dropdown.Button
+          className={styles.batchAction}
+          type={type}
+          disabled={disabled || loading}
+          size={size}
+          danger={danger}
           menu={{
             items: items,
             onClick: (parameters) => onSelect && onSelect(parameters.key),
           }}
-          className={styles.batchAction}
-          disabled={disabled}
+          onClick={onClick}
         >
-          <Button style={style}>
-            {children} {iconBuilder.down()}
-          </Button>
-        </Dropdown>
+          <IconInfo icon={icon} text={text} />
+        </Dropdown.Button>
       );
-    }}
-  </ConfigConsumer>
-);
+    }
 
-const BatchAction = ({ style, onSelect, menus = [], disabled }) => (
-  <ConfigConsumer>
-    {() => {
-      const items = menus.map((item) => {
-        return {
-          key: item.key,
-          label: item.name,
-          icon: item.icon,
-          disabled: (item.disabled || null) == null ? false : item.disabled,
-        };
-      });
+    return (
+      <Dropdown
+        menu={{
+          items: items,
+          onClick: (parameters) => onSelect && onSelect(parameters.key),
+        }}
+        className={styles.batchAction}
+        disabled={disabled || loading}
+      >
+        <Button danger={danger} size={size}>
+          <IconInfo icon={icon} text={text} />
+        </Button>
+      </Dropdown>
+    );
+  }
+}
 
-      return (
-        <Dropdown
-          menu={{
-            items: items,
-            onClick: (parameters) => onSelect && onSelect(parameters.key),
-          }}
-          className={styles.batchAction}
-          disabled={disabled}
-        >
-          <AnchorLink style={style}>{iconBuilder.ellipsis()}</AnchorLink>
-        </Dropdown>
-      );
-    }}
-  </ConfigConsumer>
-);
-
-BatchAction.Button = DropdownButton;
+BatchAction.defaultProps = {
+  type: 'default',
+  icon: null,
+  text: '按钮',
+  size: 'small',
+  danger: false,
+  menus: [],
+  onClick: null,
+  onSelect: null,
+  disabled: false,
+};
 
 export { BatchAction };