+++ packages/example-antd-management-fast-framework/src/businessComponents/Drawers/Multi/ListDrawer/BaseSimpleMultiPageDrawer/index.jsx
@@ -52,7 +52,7 @@ class BaseSimpleMultiPageDrawer extends MultiPageDrawer {
       loadApiPath: 'simple/pageList',
       // 设置默认试图模式为 table
       listViewMode: listViewConfig.viewMode.table,
-      showSelect: true,
+      showSelect: false,
       // table 显示模式行长度, 合理设置可以提升美观以及用户体验，超出可见区域将显示滚动条
       tableScrollX: 1220,
       sourceCode: '',