+++ packages/example-antd-management-fast-framework/src/businessComponents/Drawers/Multi/SelectDrawer/SimpleMultiPageMultiSelectDrawer/index.jsx
@@ -1,31 +1,39 @@
 import { connect } from 'easy-soft-dva';
 
+import { logTemplate } from 'antd-management-fast-common';
 import { switchControlAssist } from 'antd-management-fast-framework';
 
-import BaseSimpleMultiPageSelectDrawer from '../BaseSimpleMultiPageSelectDrawer';
+import { BaseSimpleMultiPageSelectDrawer } from '../BaseSimpleMultiPageSelectDrawer';
 
 import { code } from './codeSource';
 
 // 显隐控制标记, 必须设置, 标记需要全局唯一
-const visibleFlag = '771db54fe1f84b22b8c5ab869d67df21';
+const visibleFlag = '50456cfa3e5f4ee8b3fa6ff716c19553';
 
 @connect(({ simple, schedulingControl }) => ({
   simple,
   schedulingControl,
 }))
 class SimpleMultiPageMultiSelectDrawer extends BaseSimpleMultiPageSelectDrawer {
-  static open() {
-    switchControlAssist.open(visibleFlag);
-  }
-
   constructor(properties) {
     super(properties, visibleFlag);
 
     this.state = {
       ...this.state,
+      showSelect: true,
       sourceCode: code,
     };
   }
+
+  static open() {
+    switchControlAssist.open(visibleFlag);
+  }
+
+  onBatchActionClick = (list) => {
+    logTemplate({
+      list,
+    });
+  };
 }
 
 export { SimpleMultiPageMultiSelectDrawer };