+++ packages/example-antd-management-fast-framework/src/businessComponents/Drawers/Single/SelectDrawer/Normal/SimpleSinglePageMultiSelectDrawer/index.jsx
@@ -23,6 +23,7 @@ class SimpleSinglePageMultiSelectDrawer extends BaseSimpleSinglePageSelectDrawer
 
     this.state = {
       ...this.state,
+      showSelect: true,
       sourceCode: code,
     };
   }