+++ packages/example-antd-management-fast-framework/src/businessComponents/Modals/Multi/SelectModal/SimpleMultiPageMultiSelectModal/index.jsx
@@ -1,5 +1,6 @@
 import { connect } from 'easy-soft-dva';
 
+import { logTemplate } from 'antd-management-fast-common';
 import { switchControlAssist } from 'antd-management-fast-framework';
 
 import BaseSimpleMultiPageSelectModal from '../BaseSimpleMultiPageSelectModal';
@@ -23,9 +24,16 @@ class SimpleMultiPageMultiSelectModal extends BaseSimpleMultiPageSelectModal {
 
     this.state = {
       ...this.state,
+      showSelect: true,
       sourceCode: code,
     };
   }
+
+  onBatchActionClick = (list) => {
+    logTemplate({
+      list,
+    });
+  };
 }
 
 export { SimpleMultiPageMultiSelectModal };