+++ packages/antd-management-fast-common/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 2.6.62
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-common@2.6.62
+
+## 2.6.61
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-common@2.6.61
+
 ## 2.6.60
 
 ### Patch Changes